[block-stm] Move pre-writes to the executor traits
pre_write_values is a VM concern (the value is produced in the VM's
speculative representation), not a transaction property. Move it from
BlockExecutableTransaction to TransactionExecutor / ExecutorTask, with
the timestamp pre-write logic implemented by AptosExecutorTask.
Prepares for removing the value types from the transaction trait.
