+++ types/src/transaction/signature_verified_transaction.rs
@@ -4,7 +4,6 @@
 use crate::{
     block_executor::speculative_value::ValueWithLayout,
     state_store::state_key::StateKey,
-    timestamp::TimestampResource,
     transaction::{
         BlockEndInfo, BlockExecutableTransaction, FeeDistribution, SignedTransaction,
         TBlockEndInfoExt, Transaction,
@@ -135,40 +134,6 @@ impl BlockExecutableTransaction for SignatureVerifiedTransaction {
         Transaction::block_epilogue_v2(block_id, block_end_info, fee_distribution, to_make_hot)
             .into()
     }
-
-    fn pre_write_values(&self) -> Vec<(Self::Key, Self::SpeculativeValue)> {
-        let timestamp = match self {
-            SignatureVerifiedTransaction::Valid(Transaction::BlockMetadataExt(metadata_txn)) => {
-                Some(metadata_txn.timestamp_usecs())
-            },
-            SignatureVerifiedTransaction::Valid(Transaction::BlockMetadata(metadata_txn)) => {
-                Some(metadata_txn.timestamp_usecs())
-            },
-            _ => None,
-        };
-
-        match timestamp {
-            Some(ts) => {
-                // Use typed StateKey creation to avoid string parsing.
-                // These unwraps are safe: TimestampResource is a valid MoveResource type,
-                // and u64 serialization via BCS cannot fail.
-                let state_key = StateKey::resource_typed::<TimestampResource>(&AccountAddress::ONE)
-                    .expect("TimestampResource is a valid MoveResource");
-                let value = WriteOp::legacy_modification(
-                    bcs::to_bytes(&ts)
-                        .expect("u64 BCS serialization cannot fail")
-                        .into(),
-                );
-                // The timestamp resource has no delayed fields, so the pre-written value
-                // is already in its exchanged form with no layout.
-                vec![(
-                    state_key,
-                    ValueWithLayout::Exchanged(triomphe::Arc::new(value), None),
-                )]
-            },
-            None => vec![],
-        }
-    }
 }
 
 impl From<Transaction> for SignatureVerifiedTransaction {
@@ -207,51 +172,3 @@ impl TransactionProvider for Transaction {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::block_metadata::BlockMetadata;
-    use aptos_crypto::HashValue;
-
-    #[test]
-    fn test_pre_write_values_for_block_metadata() {
-        let timestamp_usecs = 1234567890u64;
-        let block_metadata = BlockMetadata::new(
-            HashValue::zero(),
-            1, // epoch
-            1, // round
-            AccountAddress::ONE,
-            vec![], // previous_block_votes_bitvec
-            vec![], // failed_proposer_indices
-            timestamp_usecs,
-        );
-
-        let txn = SignatureVerifiedTransaction::Valid(Transaction::BlockMetadata(block_metadata));
-        let pre_write_values = txn.pre_write_values();
-
-        // Should return exactly one pre-write entry for the timestamp
-        assert_eq!(pre_write_values.len(), 1);
-
-        let (state_key, value) = &pre_write_values[0];
-
-        // Verify the state key is for the timestamp resource
-        let expected_state_key =
-            StateKey::resource_typed::<TimestampResource>(&AccountAddress::ONE)
-                .expect("TimestampResource is a valid MoveResource");
-        assert_eq!(state_key, &expected_state_key);
-
-        // Verify the value is the serialized timestamp, in the exchanged form
-        // with no layout.
-        let expected_value = bcs::to_bytes(&timestamp_usecs).unwrap();
-        assert!(matches!(value, ValueWithLayout::Exchanged(_, None)));
-        assert_eq!(value.extract_value().bytes(), Some(&expected_value.into()));
-    }
-
-    #[test]
-    fn test_pre_write_values_for_user_transaction_returns_empty() {
-        // For non-block-metadata transactions, pre_write_values should return empty
-        let state_checkpoint_txn =
-            SignatureVerifiedTransaction::Valid(Transaction::StateCheckpoint(HashValue::zero()));
-        assert!(state_checkpoint_txn.pre_write_values().is_empty());
-    }
-}
